modelfitting OOP left only
